Fixed this to ensure that migrations work with sqlite, which is great for testing databases.
Added the ability from the command line to create seed the test database from the static data present in the CSVs in test.

Added a template for the .flaskenv file and removed the static one. Added to .gitignore so that people could set their secret stuff in there without checking in.

Updated the readme with a new procedure that should get people up and running quickly.