Remove pipeline-search side-effects
GitHub Issue #95 points out that pipeline cache updating
is currently done in odd places. In particular, it's done
by a background thread spawned by the pipeline-search endpoint
if either:

1. The cache directory doesn't exist
2. The cache directory is more than 5 minutes old

PR #103 fixes the bug which was preventing the
`flask update_pipeline_data` command from populating the cache,
which means the updating can be done on a periodic-basis from
cron now instead of spawning threads with side-effects when
searching for pipelines.